requirements.txt
Django==4.2
django-filter==23.1
djangorestframework==3.14.0
psycopg2-binary==2.9.6
ipython
requests==2.28.2
drf_yasg==1.21.5